refactor: remove redundant .images() structure
