+++ core/src/main/java/org/openstack4j/openstack/image/v2/domain/GlanceMemberCreate.java
@@ -1,4 +1,4 @@
-package org.openstack4j.openstack.image.v2;
+package org.openstack4j.openstack.image.v2.domain;
 
 import org.openstack4j.model.image.v2.MemberCreate;
 import org.openstack4j.model.image.v2.builder.MemberCreateBuilder;
